♻️refactor: 레이아웃 컴포넌트 디렉토리 구조 수정
